% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_geo_median.R
\name{st_geo_median_inner}
\alias{st_geo_median_inner}
\title{Calculate geometric medians from a matrix}
\usage{
st_geo_median_inner(P, tol = 1e-07, maxiter = 200)
}
\arguments{
\item{P}{A matrix of points whose geometric median will be calculated.}

\item{tol}{The tolerance}

\item{maxiter}{The maximal number of iterations}
}
\value{
A list containing the geometric median, distance,
tolerance and number of iterations before arriving at the point.
}
\description{
This function is a vectorized version of the function 'geo_median'
from the pracma package. It should give exactly the same geometric
median and sum of distances.
}
\examples{
P <- matrix(c(0, 1, 2, 0, 1.5, 2), 3, 2)
geometric_median_point <- st_geo_median_inner(P)
}
