% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_network_distance.R
\name{st_network_distance}
\alias{st_network_distance}
\title{This function calculates street distances. The user can enter a dodgr graph, else it will automatically draw one from Openstreetmap.}
\usage{
st_network_distance(
  data,
  to_data = NULL,
  index = NULL,
  placename = NULL,
  transport_mode = NULL,
  graph = NULL,
  calculate_times = FALSE
)
}
\arguments{
\item{data}{A sf dataframe}

\item{to_data}{A second sf dataframe. The function calculates the distances from all points in the first to all points in the second dataset.}

\item{index}{The name of a column of the second dataframe. Its values are used to name the distances if it is supplied. Else, distances will be numbered.}

\item{placename}{If no graph is supplied, the user has to enter a place whose street net should be used for distance calculations.}

\item{transport_mode}{If no graph is supplied, the user has to enter a transport mode. Allowed values are “bicycle”, “foot”, “goods”, “hgv”, “horse”, “moped”, “motorcar”, “motorcycle”, “psv”, or “wheelchair”.}

\item{graph}{A dodgr graph. It can be supplied and used for distance calculation, or the user supplies a place name and a mode of transport.}

\item{calculate_times}{Decide if travel times should be computed (defaults to FALSE). If FALSE, the function computes spatial distances instead.}
}
\value{
The sf dataframe 'data' with the calculated distances as additional columns.
}
\description{
This function calculates street distances. The user can enter a dodgr graph, else it will automatically draw one from Openstreetmap.
}
\examples{
# Create example graph:
graph <- dodgr::weight_streetnet(dodgr::hampi, "bicycle")
pts <- sf::st_sfc(sf::st_point(c(76.47398, 15.330)), sf::st_point(c(76.47398, 15.150)))
pts <- sf::st_as_sf(pts)
sf::st_crs(pts) <- "EPSG:3857"
pts2 <- pts[1, ]
result <- st_network_distance(data = pts, to = pts2, graph = graph)
}
